requirements/tests.txt
# Additional dependencies for the test suite

# Indirect dependencies first, exact versions for consistency

# django-nose
nose==1.3.4

# tox
py==1.4.26
virtualenv==12.0.5  # tox

# And now the direct dependencies

# For code coverage statistics generation
coverage==3.7.1

# Our testing framework
django-nose==1.3

# For managing test environments
tox==1.8.1

# Testing Yappi profile backend
yappi==0.94
